fix norme ex0

+++ C 04/Exercice 00/main.c
@@ -6,7 +6,7 @@
 /*   By: tchoquet <tchoquet@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/20 19:19:38 by tchoquet          #+#    #+#             */
-/*   Updated: 2023/02/20 19:55:01 by tchoquet         ###   ########.fr       */
+/*   Updated: 2023/02/20 20:53:17 by tchoquet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,7 @@ int	ft_strlen(char *str);
 
 int main()
 {
+	
 	char *testStr[] = {
 		"Hello World !",
 		"abcdefghijklmnopqrstuvwxyz",
@@ -25,7 +26,7 @@ int main()
 		"\t\v\t",
 	};
 
-	for (size_t i = 0; i < 6; i++)
+	for (int i = 0; i < 6; i++)
 	{
 		printf("Test with : '%s'\n", testStr[i]);
 		printf("ft_strlen : %d\n\n", ft_strlen(testStr[i]));
